+++ CHANGES.md
@@ -1,5 +1,6 @@
 **Added:**
-* deep copy of stringtable instances
+* string table editor search: can search by text value
+* redid language selection for string table editor (wanted a way to show what languages are available quickly)
 
 **Changed:**
 * none